Supporting other file extensions for the index page as well.
Coleslaw is able to write the blog content to xhtml file as well, using the
configuration setting :page-ext. But it did ignore the file extension when
creating the index page link responsible for the default landing page. With this
fix the file extension gets honored. E.g. when "xhtml" is configured, coleslaw
will generate a link from "index.xhtml" to "1.xhtml".