Add Djula theme engine plugin
There is some friction between djula templates and Coleslaw themes because the
concepts of themes in Coleslaw are strongly tied to cl-closure. Two things in
particular:

1. Templates are named by symbols. Themes have a name under
   namespaces. This is a straighforward map to CL's symbol names and
   packages.
2. Theme functions take a plist as a context, injected data
   in cl-closure terms.

In Djula on the other hand

1. Templates are named by strings or pathnames and are looked up in a
   search path
2. The RENDER-TEMPLATE* function uses keyword arguments,
   instead of a plist.

To bridge over the second difference we use the spreadable argument list
designator of APPLY

To bridge over the first difference we implement a system similar to
cl-closure, where a symbol maps to to the function to be called when
rendering the template. This mapping is stored in *TEMPLATE-STORE*.