Support extensionless sites via nil page-ext rather than index-ext.
Index-ext was a bit of a kludge used to support symlinking the index.html rather
than due to any real need for indexes to have different extensions than other
documents. Instead of an empty string to indicate no extension, we will use nil.
We hardcode "index.html" and use the first numeric-index's page-url to not make
assumptions about the user's routing scheme.